requirements.txt
aes==1.0.0
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
distlib==0.3.4
Django==3.2.12
django-appconf==1.0.5
django-crypto==0.2.0
django-cryptography==1.0
filelock==3.6.0
idna==3.3
importlib-metadata==4.11.2
Naked==0.1.31
pipenv==2022.1.8
platformdirs==2.5.1
psycopg2==2.9.3
pycparser==2.21
pycryptodome==3.14.1
pytz==2021.3
PyYAML==6.0
requests==2.27.1
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.8
virtualenv==20.13.2
virtualenv-clone==0.5.7
zipp==3.7.0
